apply code guidelines, remove unused code, show alert and inputs as readonly when is not a maintainer, improve avatar help message to include desirable size when upload an image after crop the image
