+++ client/src/components/form-field/FormGeneratorImageInput.js
@@ -46,6 +46,7 @@ import {
   InputLabel,
 } from "../formlabels/FormLabels";
 import ImageEditor, { CARD_IMAGE_DIMENSIONS } from "../imageEditor/ImageEditor";
+import { DESIRABLE_FINAL_IMAGE_SIZE } from "../../project/new/components/NewProjectAvatar";
 
 function userInputOption(options) {
   let userInput = options.find((o) => o[Prop.STOCK] === false);
@@ -258,21 +259,22 @@ function ImageContentInputMode({ name, modes, mode, setMode, onClick, color }) {
 }
 
 function ImageContentInput({
-  name,
-  value,
-  placeholder,
-  modes,
-  setInputs,
+  color,
+  disabled,
+  expectedFinalSize = DESIRABLE_FINAL_IMAGE_SIZE,
+  format,
   help,
   maxSize,
-  format,
-  disabled,
+  modes,
+  name,
   options,
+  placeholder,
   readOnly,
-  color,
-  sizeAlert,
-  setSizeAlert,
+  setInputs,
   setOriginalImageInput,
+  setSizeAlert,
+  sizeAlert,
+  value,
 }) {
   const [mode, setMode] = useState(modes[0]);
   const fileInput = useRef(null);
@@ -302,7 +304,8 @@ function ImageContentInput({
           Select an image file (max size {formatBytes(maxSize)}).
           <br />
           Images will be cropped to {CARD_IMAGE_DIMENSIONS.width}px &times;{" "}
-          {CARD_IMAGE_DIMENSIONS.height}px.
+          {CARD_IMAGE_DIMENSIONS.height}px to achieve the desirable size{" "}
+          {formatBytes(expectedFinalSize)}
         </span>
       );
     }
@@ -404,19 +407,20 @@ function ImageContentInput({
  */
 function ImageInput(props) {
   const {
-    name,
-    label,
-    value,
     alert,
-    modes,
-    setInputs,
-    help,
-    maxSize,
-    format = "image/*",
     disabled = false,
-    required = false,
-    submitting,
+    expectedFinalSize,
+    format = "image/*",
+    help,
     includeRequiredLabel,
+    label,
+    maxSize,
+    modes,
+    name,
+    required = false,
+    setInputs,
+    submitting,
+    value,
   } = props;
   const [sizeAlert, setSizeAlert] = useState(null);
   const [originalImageInput, setOriginalImageInput] = useState(null);
@@ -434,19 +438,20 @@ function ImageInput(props) {
   const contentImage = disabled ? null : (
     <div className="flex-grow-1">
       <ImageContentInput
-        name={name}
-        value={selected}
-        setInputs={setInputs}
+        disabled={disabled}
+        expectedFinalSize={expectedFinalSize}
+        format={format}
         help={help}
         maxSize={maxSize}
-        readOnly={submitting}
-        disabled={disabled}
-        options={options}
         modes={allowedModes}
-        format={format}
-        sizeAlert={sizeAlert}
-        setSizeAlert={setSizeAlert}
+        name={name}
+        options={options}
+        readOnly={submitting}
+        setInputs={setInputs}
         setOriginalImageInput={setOriginalImageInput}
+        setSizeAlert={setSizeAlert}
+        sizeAlert={sizeAlert}
+        value={selected}
       />
       {alert && <ErrorLabel text={alert} />}
     </div>
