+++ client/src/project/new/components/NewProjectAvatar.tsx
@@ -30,6 +30,7 @@ import { ImageInputMode } from "../../../components/form-field/FormGeneratorImag
 
 // 3 MB -- GitLab has a 200kB limit, but a 3 MB file should be small enough after cropping
 const PROJECT_AVATAR_MAX_SIZE = 3 * 1024 * 1024;
+const DESIRABLE_FINAL_IMAGE_SIZE = 200 * 1024;
 
 type ArtificialEventTargetValue = {
   options: { FILE: File }[];
@@ -68,6 +69,7 @@ function NewProjectAvatar({ onAvatarChange }: NewProjectAvatarProps) {
         value={value}
         help={null}
         maxSize={PROJECT_AVATAR_MAX_SIZE}
+        expectedFinalSize={DESIRABLE_FINAL_IMAGE_SIZE}
         alert={alert}
         modes={[ImageInputMode.FILE]}
         format="image/png,image/jpeg,image/gif,image/tiff"
@@ -80,4 +82,4 @@ function NewProjectAvatar({ onAvatarChange }: NewProjectAvatarProps) {
 }
 
 export default NewProjectAvatar;
-export { PROJECT_AVATAR_MAX_SIZE };
+export { PROJECT_AVATAR_MAX_SIZE, DESIRABLE_FINAL_IMAGE_SIZE };
