--- client/src/project/settings/ProjectSettings.present.js
@@ -1,47 +1,0 @@
-/*!
- * Copyright 2021 - Swiss Data Science Center (SDSC)
- * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
- * Eidgenössische Technische Hochschule Zürich (ETHZ).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- *  renku-ui
- *
- *  ProjectSettings.present.js
- *  Project settings presentational components.
- */
-
-import { Nav, NavItem } from "reactstrap";
-import { RenkuNavLink } from "../../components/RenkuNavLink";
-
-//** Navigation **//
-
-function ProjectSettingsNav(props) {
-  return (
-    <Nav
-      className="flex-column nav-light nav-pills-underline"
-      data-cy="settings-navbar"
-    >
-      <NavItem>
-        <RenkuNavLink to={props.settingsUrl} title="General" />
-      </NavItem>
-      <NavItem>
-        <RenkuNavLink to={props.settingsSessionsUrl} title="Sessions" />
-      </NavItem>
-    </Nav>
-  );
-}
-
-export { ProjectSettingsNav };
